requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
shap
matplotlib
requests
python-dotenv